add: updates to add before deploying to prod

+++ client/components/ProductsEdit/index.tsx
@@ -117,7 +117,7 @@ const ProductsEdit: React.FC<ProductsEditProps> = ({ pageId }) => {
                 onChange={e =>
                   setDetails({
                     ...productDetails,
-                    price: parseInt(e.target.value),
+                    price: parseFloat(e.target.value),
                   })
                 }
               ></Form.Control>
